move coastlines, lakes and disputed borders into component
